fix import progressbar

+++ workers/abstract_worker.py
@@ -10,7 +10,7 @@ class AbstractWorker(QtCore.QObject):
     # available signals to be used in the concrete worker
     finished = QtCore.pyqtSignal(object)
     error = QtCore.pyqtSignal(Exception, str)
-    progress = QtCore.pyqtSignal(float)
+    progress = QtCore.pyqtSignal(int)
     toggle_show_progress = QtCore.pyqtSignal(bool)
     set_message = QtCore.pyqtSignal(str)
     set_log_message = QtCore.pyqtSignal(str)
